Integrate most tools with evals
