--- ADBench_Data/BA/ba1_n49_m7776_p31843.txt
@@ -1,5 +1,0 @@
-49 7776 31843
--0.758453 -1.109613 -0.845551 34.556073 39.676747 53.881673 419.194514 5.864426 -8.518870 0.087812 0.002739 
-7.203245 0.001144 3.023326 
-0.417022 
-271.760969 834.209256
