+++ shared/gmm_data.py
@@ -5,8 +5,6 @@ from dataclasses import dataclass, field
 import numpy as np
 
 from defs import Wishart
-from output_utils import save_value_to_file, objective_file_name,\
-                            save_vector_to_file, jacobian_file_name
 
 @dataclass
 class GMMInput:
@@ -20,22 +18,6 @@ class GMMInput:
 class GMMOutput:
     objective: np.float64 = 0.0
     gradient: np.ndarray = field(default_factory=lambda: np.empty(0, dtype=np.float64))
-
-    def save_output_to_file(
-        self,
-        output_prefix,
-        input_basename,
-        module_basename
-    ):
-        save_value_to_file(
-            objective_file_name(output_prefix, input_basename, module_basename),
-            self.objective
-        )
-
-        save_vector_to_file(
-            jacobian_file_name(output_prefix, input_basename, module_basename),
-            self.gradient
-        )
 
 
 @dataclass
