+++ tools/diffsharp/Dockerfile
@@ -6,8 +6,7 @@ RUN dotnet new console -lang "F#"
 RUN dotnet add package DiffSharp-lite
 RUN dotnet add package FSharp.Data
 
-COPY tools/diffsharp/functions.fsx .
-COPY tools/diffsharp/run.fsx .
+COPY tools/diffsharp/ .
 
 # run f# through dotnet and f# interactive
 ENTRYPOINT ["dotnet", "fsi", "run.fsx"]
