+++ tools/diffsharp/run.fsx
@@ -1,20 +1,20 @@
 #r "nuget: FSharp.Data"
 #r "nuget: DiffSharp-lite"
-#load "functions.fsx"
+#load "modules.fsx"
 
 open DiffSharp
 open FSharp.Data
 open FSharp.Data.JsonExtensions
 open System
 open System.Diagnostics
-open functions
+open modules
 
 let run (pars: JsonValue) =
     let arg = dsharp.tensor (pars?input.AsFloat())
     let name = pars?name.AsString()
-    let func = if name = "square" then square else double
+    let moduleName = pars.GetProperty("module").AsString()
     let stopwatch = Stopwatch.StartNew()
-    let result = func arg
+    let result = runModule moduleName name arg
     stopwatch.Stop()
     (float result, decimal stopwatch.ElapsedTicks)
 
@@ -28,8 +28,14 @@ let createJsonData message =
             [| ("id", id)
                ("output", JsonValue.Float result)
                ("nanoseconds", JsonValue.Record [| ("evaluate", JsonValue.Number time) |]) |]
+    elif message?kind.AsString() = "define" then
+        let moduleName = message.GetProperty("module").AsString()
+        let success = moduleExists moduleName
+        JsonValue.Record [| ("id", id)
+                            ("succes", JsonValue.Boolean success) |]
     else
         JsonValue.Record [| ("id", id) |]
+
 
 
 assert (Stopwatch.Frequency = 1000000000L) //Ensure one tick is one nanosecond
