+++ tools/jax/Dockerfile
@@ -1,7 +1,6 @@
 FROM python:3.11-slim
 RUN pip install jax jaxlib
 WORKDIR /home/gradbench
-COPY tools/jax/jax_functions.py .
-COPY tools/jax/jax_run.py .
-ENTRYPOINT ["python", "/home/gradbench/jax_run.py"]
+COPY tools/jax/ .
+ENTRYPOINT ["python", "/home/gradbench/run.py"]
 LABEL org.opencontainers.image.source https://github.com/gradbench/gradbench
