+++ tools/jax/run.py
@@ -8,8 +8,8 @@ import jax.numpy as jnp
 # from jax import grad, jit
 
 
-def resolve(name):
-    functions = import_module("jax_functions")
+def resolve(module, name):
+    functions = import_module(module)
     return getattr(functions, name)
 
 
@@ -18,7 +18,7 @@ def tensor(x):
 
 
 def run(params):
-    func = resolve(params["name"])
+    func = resolve(params["module"],params["name"])
     arg = tensor(params["input"])
 
     # jfunc = jit(func)
@@ -36,7 +36,15 @@ def main():
         response = {}
         if message["kind"] == "evaluate":
             response = run(message)
-        print(json.dumps({"id": message["id"]} | response), flush=True)
+            print(json.dumps({"id": message["id"]} | response), flush=True)
+        elif message["kind"] == "define":
+            try:
+                import_module(message["module"])
+                print(json.dumps({"id": message["id"], "success": True}), flush=True)
+            except:
+                print(json.dumps({"id": message["id"], "success": False}), flush=True)
+        else:
+             print(json.dumps({"id": message["id"]}), flush=True)
 
 
 if __name__ == "__main__":
