+++ tools/pytorch/gradbench.py
@@ -1,0 +1,7 @@
+def square(x):
+    return x * x
+
+def double(x):
+    y = square(x)
+    y.backward()
+    return x.grad
