+++ tools/tapenade/Dockerfile
@@ -16,11 +16,7 @@ RUN tar -xvzf tapenade_3.16.tar
 # ensure Tapenade is accessible
 RUN ln -s /home/gradbench/tapenade_3.16/bin/tapenade /usr/local/bin/tapenade
 
-COPY tools/tapenade/functions.c .
-COPY tools/tapenade/run_deriv.c .
-COPY tools/tapenade/run_origin.c .
-COPY tools/tapenade/run_shells.py .
-COPY tools/tapenade/tapenade_run.py .
+COPY tools/tapenade/ .
 
-ENTRYPOINT ["python3", "/home/gradbench/tapenade_run.py"]
+ENTRYPOINT ["python3", "/home/gradbench/run.py"]
 LABEL org.opencontainers.image.source https://github.com/gradbench/gradbench
