+++ tools/tensorflow/Dockerfile
@@ -7,7 +7,8 @@ RUN pip install h5py
 RUN pip install tensorflow
 
 WORKDIR /home/gradbench
-COPY tools/tensorflow/tensor_functions.py .
-COPY tools/tensorflow/tensor_run.py .
-ENTRYPOINT ["python", "/home/gradbench/tensor_run.py"]
+
+COPY tools/tensorflow/ .
+
+ENTRYPOINT ["python", "/home/gradbench/run.py"]
 LABEL org.opencontainers.image.source https://github.com/gradbench/gradbench
