+++ tools/tensorflow/run.py
@@ -6,8 +6,8 @@ from importlib import import_module
 import tensorflow as tf
 
 
-def resolve(name):
-    functions = import_module("tensor_functions")
+def resolve(module, name):
+    functions = import_module(module)
     return getattr(functions, name)
 
 
@@ -16,7 +16,7 @@ def tensor(x):
 
 
 def run(params):
-    func = resolve(params["name"])
+    func = resolve(params["module"],params["name"])
     arg = tensor(params["input"])
     start = time.perf_counter_ns()
     ret = func(arg)
@@ -31,7 +31,15 @@ def main():
         response = {}
         if message["kind"] == "evaluate":
             response = run(message)
-        print(json.dumps({"id": message["id"]} | response), flush=True)
+            print(json.dumps({"id": message["id"]} | response), flush=True)
+        elif message["kind"] == "define":
+            try:
+                import_module(message["module"])
+                print(json.dumps({"id": message["id"], "success": True}), flush=True)
+            except:
+                print(json.dumps({"id": message["id"], "success": False}), flush=True)
+        else:
+             print(json.dumps({"id": message["id"]}), flush=True)
 
 
 if __name__ == "__main__":
